+++ app/(mainpage)/feed/page.jsx
@@ -1,0 +1,8 @@
+import React from "react";
+
+import MyFeed from '@components/MyFeed';
+
+
+export default function feed() {
+  return <div>feed</div>;
+}
