+++ app/(mainpage)/topic/[topic_name]/page.jsx
