--- app/[blog_name]/manage/newpost/layout.jsx
@@ -1,5 +1,0 @@
-import React from "react";
-
-export default function NewpostLayout({ children }) {
-  return <section>{children}</section>;
-}
