+++ app/[blog_name]/newpost/layout.jsx
@@ -1,0 +1,5 @@
+import React from "react";
+
+export default function NewpostLayout({ children }) {
+  return <section>{children}</section>;
+}
