--- app/feed/page.jsx
@@ -1,8 +1,0 @@
-import React from "react";
-
-import MyFeed from '@components/MyFeed';
-
-
-export default function feed() {
-  return <div>feed</div>;
-}
