--- app/topic/culture/page.jsx
@@ -1,3 +1,0 @@
-export default function culture() {
-  return <div>culture</div>;
-}
