--- app/topic/it/page.jsx
@@ -1,3 +1,0 @@
-export default function it() {
-  return <div>it</div>;
-}
