--- app/topic/life/page.jsx
@@ -1,3 +1,0 @@
-export default function life() {
-  return <div>life</div>;
-}
