--- app/topic/sports/page.jsx
@@ -1,3 +1,0 @@
-export default function sports() {
-  return <div>sports</div>;
-}
