--- app/topic/travel/page.jsx
@@ -1,3 +1,0 @@
-export default function travel() {
-  return <div>travel</div>;
-}
